% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateProjectDirectory.R
\name{CreateProjectDirectory}
\alias{CreateProjectDirectory}
\title{Create project directory structure}
\usage{
CreateProjectDirectory(name, template)
}
\arguments{
\item{name}{Name of project}

\item{template}{Template name. Must refer to one of the json files stored in data/templates.}
}
\description{
Create project directory structure
}
\details{
Functions creates a project folder with the specified name in the
current working directory. The directory structure of the folder is specified
by the provided template. Function halt with an error if a folder with the
specified project name already exists.
}
