% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadTemplate.R
\name{LoadTemplate}
\alias{LoadTemplate}
\title{Import project directory template}
\usage{
LoadTemplate(template)
}
\arguments{
\item{template}{Template name. Must refer to one of the json files stored in data/templates.}
}
\value{
An object of class data.tree
}
\description{
Import project directory template
}
\details{
Function imports a json file describing a project directory, and
creates a tree object.
}
